AD-HOC feat (*): Initial Commit
This commit managers user access to a machine. It's simple, and
optionally allows specifying users that allowed to access root
privileges through "sudo".

At this time, it's simply a stub, slightly better than per project
roles. However, in future it may be expanded at which point it can be
updated in the project.