feat (user removal): adds more flags for user removal
Currently we loop over custom user attributes and map them
to user role of ansible. for removing a user it can be benfitial
if we can have remove and force attributes.

These attributes can help to remove user's home folder and related
files which are not removed by just stating `state=absent`